Clean up and improve tests
